+++ app/models/activity.rb
@@ -63,6 +63,7 @@ class Activity < ApplicationRecord
 
   scope :funds, -> { where(level: :fund) }
   scope :programmes, -> { where(level: :programme) }
+  scope :publishable_to_iati?, -> { where(form_state: :complete) }
 
   def valid?(context = nil)
     context = VALIDATION_STEPS if context.nil? && form_steps_completed?